Allow moving a layer to the bottom-most spot
Before it wouldn't let you do it due to an off-by-one error. If a layer
is moved down (to a higher index), the code would decrement the target
index by one to account for the indexes shifting by one, but only did so
*after* checking if the operation was okay. Now the check uses the
correct target index.

This causes no breakage when drawing together with folks that don't have
this fix, since the command that's sent always transmits the entire
resulting layer order, rather than just the move operation.