Rework animation system into a real timeline
Throws away the previous manual timeline and replaces it with something
more akin to what you see in other programs that support animation.

The timeline now consists of tracks and key frames. Each key frame
either refers to a layer id or is blank. The visibility of layers within
a key frame can be filtered, allowing re-use of the same group in
several different configurations without having to copy stuff around or
split it up into multiple tracks. When rendering a frame, the tracks act
as the layer stack, using the closest key frames to render the canvas.
Onion skins can be enabled per-track.

This is still somewhat different from the way other programs do it,
since they conflate layers and tracks and basically give you multiple
versions of every layer. But I don't like that because I have to copy
stuff around a bunch, so I consider this way of doing it superior.

There's still further work to do on it, but at this point it's pretty
much fully functional, so I picked it as the cut-off point to not
balloon this commit even larger.