Rework Flipbook to be more comfortable to use
It now has a refresh button so that you don't have to keep opening and
closing it. Re-issuing the Flipbook shortcut will also no longer open a
second one, instead it will just refresh the currently open one. The
speed can now be adjusted using a speed slider, rather than having the
confusing FPS spinner in addition to the actual FPS set in the timeline.