Patch androiddeployqt for proper APK signing
The Qt5 androiddeployqt expects you to give the keystore password on the
command line, which sucks because it will get dumped into logs and into
build scripts. Additionally, Qt5 makes it really hard to actually give
additional arguments to androiddeployqt. So instead, this patch
backports Qt6 functionality and allows passing that stuff through
environment variables instead, which is safer and saner.